Rename cliplugin directory from kcp-edge to kubestellar
