+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "go.sum|docs/package-lock.json|cmd/kube-watchall/go.sum|docs/config.toml|docs/resources/_gen|docs/static|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2023-06-01T21:04:21Z",
+  "generated_at": "2023-06-11T03:51:35Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -87,7 +87,7 @@
         "verified_result": null
       }
     ],
-    "pkg/cliplugins/kcp-edge/syncer-gen/edgesync.go": [
+    "pkg/cliplugins/kubestellar/syncer-gen/edgesync.go": [
       {
         "hashed_secret": "4f151c44a6783667af6cedaa3d12560f8af2f38a",
         "is_secret": false,
@@ -105,12 +105,12 @@
         "verified_result": null
       }
     ],
-    "test/e2e/framework/edgesyncer.go": [
+    "test/e2e/framework/kubestellar_syncer.go": [
       {
         "hashed_secret": "4d55af37dbbb6a42088d917caa1ca25428ec42c9",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 358,
+        "line_number": 353,
         "type": "Secret Keyword",
         "verified_result": null
       }
