+++ cmd/kubectl-kubestellar-syncer_gen/main.go
@@ -27,7 +27,7 @@ import (
 	"k8s.io/component-base/version"
 	"k8s.io/klog/v2"
 
-	plugin "github.com/kcp-dev/edge-mc/pkg/cliplugins/kcp-edge/syncer-gen"
+	plugin "github.com/kcp-dev/edge-mc/pkg/cliplugins/kubestellar/syncer-gen"
 )
 
 var (
