+++ pkg/cliplugins/kubestellar/syncer-gen/edgesync.go
