+++ pkg/cliplugins/kubestellar/syncer-gen/edgesync_test.go
