+++ pkg/cliplugins/kubestellar/syncer-gen/kubestellar-syncer.yaml
